Splitting into utils and script
